Fixed audio volume

+++ obs-studio-server/source/osn-input.cpp
@@ -355,7 +355,7 @@ void osn::Input::SetVolume(
 		PRETTY_ERROR_RETURN(ErrorCode::InvalidReference, "Input reference is not valid.");
 	}
 
-	obs_source_set_volume(input, (float_t)args[1].value_union.fp64);
+	obs_source_set_volume(input, (float_t)args[1].value_union.fp32);
 
 	rval.push_back(ipc::value((uint64_t)ErrorCode::Ok));
 	rval.push_back(ipc::value(obs_source_get_volume(input)));
